Mobile Only FullScreen Touch to Cylce
Quite pleased with this one.

Mobile only touch to cycle between scenes. Will cycle up - then back down again - then back up again etc. etc.

WURFL.io added back in to cater for the device detection client side. 

Next, Previous and Cycle function are 'clean' - and maintain a correct index. Additionally - to allow for the cycle function,  a boolean variable of 'cycled' is created, set to 'true' when the final scene is visted and flase when the first scene is visited. If cycled = true, the function will send you backwards, 'false' and your going forward.

The is a global scope variable atm - which isn't that sensible considering it is NOT set by the next and previous functions. i.e it doesn't work nice and cleanly like the indexing - but it does the job.

Worth either localising the variable to the specific function - or expanding the next and previous functions to set this variable appropriatley. Seems worth it from a future proofing perspective - but in reality I can see any wider use for it.